refactor: Move Chat Deactivation to common
