+++ common/chat/deactivation.ts
@@ -1,0 +1,101 @@
+import moment from 'moment';
+import { markConversationAsReadOnly, sendSystemMessage, updateConversation } from './conversation';
+import { countChatParticipants, checkChatMembersAccessRights } from './helper';
+import systemMessages from './localization';
+import { TJConversation, FinishedReason, SystemMessage } from './types';
+import { prisma } from '../prisma';
+import { getLogger } from '../logger/logger';
+
+const logger = getLogger('Chat Deactivation');
+
+// one to one chats (if match) whose match was dissolved 30 days ago should be "disabled" (readonly).
+// This will allow users to continue writing for another 30 days after match disolving.
+async function isActiveMatch(id: number): Promise<boolean> {
+    const today = moment().endOf('day');
+    const match = await prisma.match.findUniqueOrThrow({ where: { id } });
+    const dissolvedAtPlus30Days = moment(match.dissolvedAt).add(30, 'days');
+    return !dissolvedAtPlus30Days.isBefore(today);
+}
+
+async function isActiveSubcourse(id: number): Promise<boolean> {
+    const today = moment().endOf('day');
+    const subcourse = await prisma.subcourse.findUniqueOrThrow({ where: { id }, include: { lecture: true } });
+    const isSubcourseCancelled = subcourse.cancelled;
+
+    if (isSubcourseCancelled) {
+        return false;
+    }
+
+    const lastLecture = subcourse.lecture.sort((a, b) => moment(a.start).milliseconds() - moment(b.start).milliseconds()).pop();
+    const lastLecturePlus30Days = moment(lastLecture.start).add(30, 'days');
+    const is30DaysBeforeToday = lastLecturePlus30Days.isBefore(today);
+    return !is30DaysBeforeToday;
+}
+
+export function isConversationReadOnly(conversation: TJConversation) {
+    const countParticipants = countChatParticipants(conversation);
+    const { readMembers } = checkChatMembersAccessRights(conversation);
+    return readMembers.length === countParticipants;
+}
+
+function isMatchConversation(conversation: TJConversation) {
+    return !!conversation.custom.match;
+}
+
+export async function deactivateConversation(conversation: TJConversation) {
+    // Prevent users from writing new messages
+    await markConversationAsReadOnly(conversation.id);
+
+    // Inform users about the deactivation
+    const systemMessage = isMatchConversation(conversation) ? SystemMessage.ONE_ON_ONE_OVER : SystemMessage.GROUP_OVER;
+    await sendSystemMessage(systemMessages.de.deactivated, conversation.id, systemMessage);
+
+    // Store the deactivation reason in the Chat, in case we need this somewhen
+    await updateConversation({
+        id: conversation.id,
+        custom: {
+            finished: isMatchConversation(conversation) ? FinishedReason.MATCH_DISSOLVED : FinishedReason.COURSE_OVER,
+        },
+    });
+
+    logger.info('Deactivated Conversation', { conversationId: conversation.id, systemMessage });
+}
+
+// Whether a Chat should be deactivated by the background job soon as the corresponding
+// match or subcourse ended some time ago
+export async function shouldMarkChatAsReadonly(conversation: TJConversation) {
+    let shouldMarkAsReadonly = true;
+
+    if (conversation.custom.subcourse) {
+        const subcourseIds: number[] = JSON.parse(conversation.custom.subcourse);
+        const allSubcoursesActive = await Promise.all(subcourseIds.map((id) => isActiveSubcourse(id)));
+        const allInactive = allSubcoursesActive.every((active) => active === false);
+        logger.info(`Conversation ${conversation.id} belongs to subcourses which are ${allInactive ? 'all inactive' : 'active'}`, {
+            conversationId: conversation.id,
+        });
+        shouldMarkAsReadonly &&= allInactive;
+    }
+
+    if (conversation.custom.prospectSubcourse) {
+        const prospectSubcourses: number[] = JSON.parse(conversation.custom.prospectSubcourse);
+        const allProspectSubcoursesActive = await Promise.all(prospectSubcourses.map((id) => isActiveSubcourse(id)));
+        const allInactive = allProspectSubcoursesActive.every((active) => active === false);
+        logger.info(`Conversation ${conversation.id} belongs to subcourses which are all ${allInactive ? 'all inactive' : 'active'}`, {
+            conversationId: conversation.id,
+        });
+        shouldMarkAsReadonly &&= allInactive;
+    }
+
+    if (conversation.custom.match) {
+        const match = JSON.parse(conversation.custom.match);
+        const matchId = match.matchId;
+        const isMatchActive = await isActiveMatch(matchId);
+        logger.info(`Conversation ${conversation.id} belongs to Match(${matchId}) which is all ${isMatchActive ? 'active' : 'inactive'}`, {
+            conversationId: conversation.id,
+            matchId,
+        });
+        shouldMarkAsReadonly &&= !isMatchActive;
+    }
+
+    return shouldMarkAsReadonly;
+}
